Moved plugin out of index
